third_party/WebKit/LayoutTests/inspector/console/console-log-custom-elements-expected.txt

CONSOLE MESSAGE: line 11: [object HTMLElement]
CONSOLE MESSAGE: line 23: [object HTMLElement]
Tests that logging custom elements uses proper formatting.

console-log-custom-elements.html:11 foo-bar
console-log-custom-elements.html:23 foo-bar2
